リファクタ: build_pdf を責務別サブモジュールへ分割する
Closes #145
