+++ crates/seiran/src/build_pdf/running.rs
@@ -1,0 +1,55 @@
+//! ヘッダー・フッターの配置仕様組み立て
+
+use layout::{RunningContentSpec, RunningMetadata, RunningSlots};
+use read_config::DocumentConfig;
+use read_style::{RunningContentStyle, Style};
+use types::Color;
+
+/// ページ数確定後のヘッダー・フッター配置仕様 [`layout::RunningContentSpec`] を組み立てる。
+///
+/// ヘッダー / フッターの各スロットは [`running_slots`] で構築し（全スロット空なら描画省略）、`skip_first`
+/// でタイトルページ（先頭ページ）への非描画を指示する。`page_numbers` のラベル解決は配置パス側が担う。
+pub(super) fn build_running_spec(
+  style: &Style,
+  document: &DocumentConfig,
+  text_width: f32,
+  page_height: f32,
+  page_numbers: Vec<(String, String)>,
+) -> RunningContentSpec {
+  return RunningContentSpec {
+    header: running_slots(&style.header, style.header.baseline_offset.to_pt(), true),
+    footer: running_slots(&style.footer, page_height - style.footer.baseline_offset.to_pt(), false),
+    metadata: RunningMetadata {
+      title: document.title.clone().unwrap_or_default(),
+      author: document.author.clone().unwrap_or_default(),
+      date: document.date.clone().unwrap_or_default(),
+    },
+    text_width,
+    page_numbers,
+    // タイトルページ（先頭ページ）にはヘッダー・フッターを描画しない
+    skip_first: style.title_page.enabled,
+  };
+}
+
+/// `RunningContentStyle` をヘッダー・フッター配置用の [`layout::RunningSlots`] に変換する。
+///
+/// 全スロットが空のリージョンは描画不要なので `None` を返し、配置パスを省略させる。
+/// `baseline_y` はベースラインのページ上端からの絶対距離（フッターは呼び出し側で換算済み）、
+/// `rule_below` は区切り線をテキストの下に置くか（ヘッダーは `true`、フッターは `false`）。
+fn running_slots(style: &RunningContentStyle, baseline_y: f32, rule_below: bool) -> Option<RunningSlots> {
+  if style.is_empty() {
+    return None;
+  }
+  return Some(RunningSlots {
+    left: style.left.clone(),
+    center: style.center.clone(),
+    right: style.right.clone(),
+    font_kind: style.font_kind,
+    font_size: style.font_size.to_pt(),
+    baseline_y,
+    rule_below,
+    rule_thickness: style.rule_thickness.to_pt(),
+    rule_gap: style.rule_gap.to_pt(),
+    rule_color: style.rule_color.map(Color::rgb),
+  });
+}
